add browser location feature to location component, not fully finished some problems with passing lat and long in url
Able to ask user for location and will get location and display in innerHTML but unable to use the data in url or routerLink to send to the movieGlu api, need to fix this and send to api and add drop-down list of locations for a user to pick from if not willing to give location. moving on with other features while stuck.