add ngx-smart-modal and style the movie component. add ifrome for movie trailer
styled the movie component to look good on mobile and loptop. downloaded ngx-smart-modal to fit the iframe into a model as it would not sit in the right postion on the page when in laptop mode. now the user will see a button on top of the page saying 'play trailer' and when clicked will be able to watch the trailer in the model.