install ng translate core / loader and extract
to get this to work you needed to download extract.translations and run npm run extract-translations to update the json files and make them work. needed to add a line of code in package.json to point to it. if not done you will get a parsing error that i was unable to find online or fix till i tryed extracting. good help here https://www.codeandweb.com/babeledit/tutorials/how-to-translate-your-angular7-app-with-ngx-translate